The plain Makefile was using -Wall, but the gyp build wasn't.  This CL turns on -Wall -Wextra and -Wno-unused in common.gypi.  This revealed a lot of warnings (and some actual bugs), all of which I fixed here.  This is pretty mindless stuff for the most part (order of intialization, missing initializers, && within ||, etc), but will allow us to build cleanly with -Wall and -Wextra (and -Werror, if we so choose).
I put defaults into switches that were missing cases.  I could put in the actual
missing enums instead if that's desired.  I could also assert on missing enums
instead of break, if that's desired.  I wasn't sure how to test the stuff in
"animator", so that should be looked at a bit more closely.

Review URL:  http://codereview.appspot.com/4547055/



git-svn-id: http://skia.googlecode.com/svn/trunk@1386 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ gpu/src/gr_unittests.cpp
@@ -68,7 +68,7 @@ static void test_bsearch() {
     for (size_t n = 0; n < GR_ARRAY_COUNT(array); n++) {
         for (size_t i = 0; i < n; i++) {
             int index = GrTBSearch<int, int>(array, n, array[i]);
-            GrAssert(index == i);
+            GrAssert(index == (int) i);
             index = GrTBSearch<int, int>(array, n, -array[i]);
             GrAssert(index < 0);
         }
